requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
Pillow==7.2.0
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1